Type Alias halo2_middleware::circuit::ExpressionMid

pub type ExpressionMid<F> = Expression<F, VarMid>;

Aliased Type

enum ExpressionMid<F> {
    Constant(F),
    Var(VarMid),
    Negated(Box<Expression<F, VarMid>, Global>),
    Sum(Box<Expression<F, VarMid>, Global>, Box<Expression<F, VarMid>, Global>),
    Product(Box<Expression<F, VarMid>, Global>, Box<Expression<F, VarMid>, Global>),
}

Variants

Constant(F)

This is a constant polynomial

Var(VarMid)

This is a variable

Negated(Box<Expression<F, VarMid>, Global>)

This is a negated polynomial

Sum(Box<Expression<F, VarMid>, Global>, Box<Expression<F, VarMid>, Global>)

This is the sum of two polynomials

Product(Box<Expression<F, VarMid>, Global>, Box<Expression<F, VarMid>, Global>)

This is the product of two polynomials

Implementations

impl<F: Field, V: Variable> Expression<F, V>

pub fn evaluate<T>( &self, constant: &impl Fn(F) -> T, var: &impl Fn(V) -> T, negated: &impl Fn(T) -> T, sum: &impl Fn(T, T) -> T, product: &impl Fn(T, T) -> T ) -> T

Evaluate the polynomial using the provided closures to perform the operations.

pub fn write_identifier<W: Write>(&self, writer: &mut W) -> Result<()>

pub fn identifier(&self) -> String

Identifier for this expression. Expressions with identical identifiers do the same calculation (but the expressions don’t need to be exactly equal in how they are composed e.g. 1 + 2 and 2 + 1 can have the same identifier).

pub fn degree(&self) -> usize

Compute the degree of this polynomial

pub fn complexity(&self) -> usize

Approximate the computational complexity of this expression.