Trait halo2_middleware::expression::Variable

pub trait Variable: Clone + Copy + Debug + Eq + PartialEq {
    // Required methods
    fn degree(&self) -> usize;
    fn write_identifier<W: Write>(&self, writer: &mut W) -> Result<()>;

    // Provided method
    fn complexity(&self) -> usize { ... }
}

Required Methods

fn degree(&self) -> usize

Degree that an expression would have if it was only this variable.

fn write_identifier<W: Write>(&self, writer: &mut W) -> Result<()>

Write an identifier of the variable. If two variables have the same identifier, they must be the same variable.

Provided Methods

fn complexity(&self) -> usize

Approximate the computational complexity an expression would have if it was only this variable.

Object Safety

This trait is not object safe.