Module halo2_middleware::circuit

Structs

Cell

A cell identifies a position in the plonkish matrix identified by a column and a row offset.

ChallengeMid

A challenge squeezed from transcript after advice columns at the phase have been committed.

ColumnMid

A column with an index and type

CompiledCircuit

This is a description of a low level Plonkish compiled circuit. Contains the Constraint System as well as the fixed columns and copy constraints information.

ConstraintSystemMid

This is a description of the circuit environment, such as the gate, column and permutation arrangements.

Gate

A Gate contains a single polynomial identity with a name as metadata.

Preprocessing

Data that needs to be preprocessed from a circuit

QueryMid

Enums

Any

An enum over the Advice, Fixed, Instance structs

VarMid

Type Aliases

ExpressionMid

GateMid